Move Credo Global inspection to namespace
